refactor: update api routes
