+++ examples/with-esm/src/routes/hello-world.get.mjs
@@ -1,0 +1,7 @@
+import { Route } from '@sapphire/plugin-api';
+
+export class UserRoute extends Route {
+	run(_request, response) {
+		response.json({ message: 'Hello World' });
+	}
+}
