+++ examples/with-esm/src/routes/index.post.mjs
@@ -1,0 +1,7 @@
+import { Route } from '@sapphire/plugin-api';
+
+export class UserRoute extends Route {
+	run(_request, response) {
+		response.json({ message: 'Landing Page!' });
+	}
+}
