+++ examples/with-javascript/src/routes/hello-world.get.js
@@ -1,0 +1,11 @@
+const { Route } = require('@sapphire/plugin-api');
+
+class UserRoute extends Route {
+	run(_request, response) {
+		response.json({ message: 'Hello World' });
+	}
+}
+
+module.exports = {
+	UserRoute
+};
