+++ examples/with-typescript-complete/src/routes/hello-world.post.ts
@@ -1,0 +1,7 @@
+import { Route } from '@sapphire/plugin-api';
+
+export class UserRoute extends Route {
+	public override run(_request: Route.Request, response: Route.Response) {
+		response.json({ message: 'Hello World' });
+	}
+}
